refactor(worker): extract model/lora download+import jobs into model_jobs.rs (sc-1636)
Move the model/lora acquisition cluster out of lib.rs into model_jobs.rs:
run_model_download_job, run_model_convert_job, run_model_import_job,
run_lora_import_job and their supporting helpers (HF CLI download,
converted-dir finalize, in-target family detection, the
DownloadFamilyCheck/FamilyMismatch reconcile types).

Same pattern: `use super::*;` (the download/source-url helpers, install
markers, target resolvers, job-lifecycle helpers, and the ApiClient all
resolve from the root/other modules unchanged), pub(crate) on the items
reached from the job dispatcher + tests, glob re-import at the root,
worker tests.rs imports repointed to super::model_jobs.
Behavior-preserving. lib.rs 4,350 -> 3,344 lines.
